refactor(handlers): extract search.ts handler from intent.ts (Stage B)
handleSAPSearch + lookup/transliteration helpers. transliterateQuery +
looksLikeFieldName re-exported (barrel-locked). 3672 tests, A1 zero-diff.
